[buffer] Extract buffer_box to separate heaaderfile
